requirements.txt
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pandas==2.1.1
numpy==1.26.0
earthengine-api==0.1.375
python-dotenv==1.0.0
PyJWT==2.8.0
